add dispatch pick detail confirm api

+++ backend/ModernWMS.WMS/Entities/ViewModels/Dispatchlist/DispatchlistViewModel.cs
@@ -295,7 +295,7 @@ namespace ModernWMS.WMS.Entities.ViewModels
         /// <summary>
         /// pick_checker_id
         /// </summary>
-        public byte pick_checker_id { get; set; } = 0;
+        public int pick_checker_id { get; set; } = 0;
 
         /// <summary>
         /// pick_checker
